# docs/swagger.yaml
openapi: 3.0.0
info:
  title: KYC Verification API
  version: 1.0.0
  description: API for KYC verification and risk assessment
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    KYCVerification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        type:
          type: string
          enum: [individual, corporate]
        status:
          type: string
          enum: [pending, in_progress, approved, rejected, expired]
        documentType:
          type: string
        documentNumber:
          type: string
        documentExpiryDate:
          type: string
          format: date
        documentFrontUrl:
          type: string
        documentBackUrl:
          type: string
        selfieUrl:
          type: string
        verificationData:
          type: object
        riskScore:
          type: number
        riskFactors:
          type: array
          items:
            type: string
        verificationNotes:
          type: string
        verifiedBy:
          type: string
          format: uuid
        verifiedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /kyc/submit:
    post:
      summary: Submit KYC verification
      tags: [KYC]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - documentType
                - documentNumber
                - documentExpiryDate
                - documentFrontUrl
                - documentBackUrl
                - selfieUrl
              properties:
                type:
                  type: string
                  enum: [individual, corporate]
                documentType:
                  type: string
                documentNumber:
                  type: string
                documentExpiryDate:
                  type: string
                  format: date
                documentFrontUrl:
                  type: string
                documentBackUrl:
                  type: string
                selfieUrl:
                  type: string
      responses:
        '201':
          description: KYC verification submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KYCVerification'
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized
...